- Kai Germaschewski: ISDN update (including Makefiles) - Jens Axboe: cdrom updates - Petr Vandrovec; Matrox G450 support - Bill Nottingham: fix FAT32 filesystems on 64-bit platforms - David Miller: sparc (and other) Makefile fixup - Andrea Arkangeli: alpha SMP TLB context fix (and cleanups) - Niels Kristian Bech Jensen: checkconfig, USB warnings - Andrew Grover: large ACPI update

+++ drivers/acpi/interpreter/amsystem.c
@@ -2,7 +2,7 @@
 /******************************************************************************
  *
  * Module Name: amsystem - Interface to OS services
- *              $Revision: 51 $
+ *              $Revision: 52 $
  *
  *****************************************************************************/
 
@@ -202,7 +202,7 @@ acpi_aml_system_acquire_mutex (
 	}
 
 	status = acpi_aml_system_wait_semaphore (obj_desc->mutex.semaphore,
-			  time_desc->number.value);
+			  (u32) time_desc->number.value);
 	return (status);
 }
 
@@ -299,7 +299,7 @@ acpi_aml_system_wait_event (
 
 	if (obj_desc) {
 		status = acpi_aml_system_wait_semaphore (obj_desc->event.semaphore,
-				  time_desc->number.value);
+				  (u32) time_desc->number.value);
 	}
